[IMP] stock_release_channel_partner_by_date: tests common
